requirements.txt
# Core Dependencies
tensorflow>=2.10.0
numpy>=1.21.0
pillow>=9.0.0
matplotlib>=3.5.0

# Data Processing
scikit-learn>=1.0.0
pandas>=1.3.0

# Progress Bars
tqdm>=4.62.0

# Text-to-Speech
pyttsx3>=2.90

# Web Interface
streamlit>=1.20.0

# Optional but Recommended
# jupyter>=1.0.0
# ipywidgets>=8.0.0